+++ packages/registration_delivery/CHANGELOG.md
@@ -1,6 +1,12 @@
+
+## 1.0.3-dev.4
+
+* Cache build fix for registration_bloc
+
 ## 1.0.3-dev.3
 
-* Pagination offset fix
+* Closed household flow change
+
 
 ## 1.0.3-dev.2